moved pages to pages directory
